backend_old/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]
